+++ packages/mock/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @graphql-tools/mock
 
+## 9.1.12
+
+### Patch Changes
+
+- Updated dependencies
+  [[`e90719b`](https://github.com/ardatan/graphql-tools/commit/e90719baf08bf1aa7b372759dd197ac27e71dc64)]:
+  - @graphql-tools/utils@11.2.2
+  - @graphql-tools/schema@10.0.38
+
+## 9.1.11
+
+### Patch Changes
+
+- Updated dependencies
+  [[`615c1a2`](https://github.com/ardatan/graphql-tools/commit/615c1a27f2a379eac74127815effc8e062bdc47c)]:
+  - @graphql-tools/utils@11.2.1
+  - @graphql-tools/schema@10.0.37
+
 ## 9.1.10
 
 ### Patch Changes